refactor: 주문 목록 조회 최적화(Slice) 및 응답 DTO 생성,           API HTTP 메서드 수정
- OrderSearchResponseDto 추가
- 무한 스크롤 통일성을 위해 조회 방식을 Slice로 변경
- http 메서드 수정 (수정 : POST -> PATCH, 조회 : POST -> GET)

+++ src/main/java/com/project/qr_order_system/dto/admin/AdminOrderSearchDto.java
@@ -18,8 +18,8 @@ public class AdminOrderSearchDto {
     private LocalDateTime startDate;
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
     private LocalDateTime endDate;
-    private Integer minPrice; // 최소 주문 금액
-    private Integer maxPrice; // 최대 주문 금액
+    private Long minPrice; // 최소 주문 금액
+    private Long maxPrice; // 최대 주문 금액
     private Long userId; // 특정 회원 조회
     private Long menuId; // 특정 메뉴 조회
